requirements.txt
python-telegram-bot==20.7
httpx==0.24.1
cachetools==5.3.0
python-dotenv==1.0.0
aiosqlite==0.19.0
structlog==23.1.0